deployment/requirements.txt
certifi
charset-normalizer
click
Flask
flatbuffers
idna
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
numpy
onnx
onnxruntime
pandas
Pillow
protobuf
python-dateutil
pytz
requests
six
# torch==1.11.0+cpu
# torchaudio==0.11.0+cpu
# torchvision==0.12.0+cpu
torch
torchvision
torchaudio
typing_extensions
urllib3
Werkzeug
zipp
gunicorn
